"Removed Integrations module and its related files and dependencies from the codebase."
