--- src/common/integrations/integrations.module.ts
@@ -1,7 +1,0 @@
-import { Global, Module } from '@nestjs/common';
-
-@Global()
-@Module({
-  imports: [],
-})
-export class IntegrationsModule {}
