+++ src/company/company.module.ts
@@ -9,6 +9,8 @@ import { BranchModule } from './branch/branch.module';
 import { BranchService } from './branch/branch.service';
 import { PermissionModule } from './permission/permission.module';
 import { DepartmentModule } from './department/department.module';
+import { SettingsModule } from './settings/settings.module';
+import { IntegrationsModule } from './integrations/integrations.module';
 
 @Module({
   imports: [
@@ -18,6 +20,8 @@ import { DepartmentModule } from './department/department.module';
     BranchModule,
     PermissionModule,
     DepartmentModule,
+    SettingsModule,
+    IntegrationsModule,
   ],
   providers: [BranchService, CompanyService, EmployeeService, JobRoleService],
   exports: [BranchModule, CompanyService, EmployeeModule, UserModule],
