+++ src/company/integrations/dto/create-integration.dto.ts
@@ -1,0 +1,1 @@
+export class CreateIntegrationDto {}
