+++ src/company/integrations/dto/update-integration.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateIntegrationDto } from './create-integration.dto';
+
+export class UpdateIntegrationDto extends PartialType(CreateIntegrationDto) {}
