+++ src/company/integrations/entities/integration.entity.ts
@@ -1,0 +1,1 @@
+export class Integration {}
