+++ src/company/integrations/integrations.controller.ts
@@ -1,0 +1,33 @@
+import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { IntegrationsService } from './integrations.service';
+
+
+@Controller('integrations')
+export class IntegrationsController {
+  constructor(private readonly integrationsService: IntegrationsService) {}
+
+  // @Post()
+  // create(@Body() createIntegrationDto: CreateIntegrationDto) {
+  //   return this.integrationsService.create(createIntegrationDto);
+  // }
+
+  // @Get()
+  // findAll() {
+  //   return this.integrationsService.findAll();
+  // }
+
+  // @Get(':id')
+  // findOne(@Param('id') id: string) {
+  //   return this.integrationsService.findOne(+id);
+  // }
+
+  // @Patch(':id')
+  // update(@Param('id') id: string, @Body() updateIntegrationDto: UpdateIntegrationDto) {
+  //   return this.integrationsService.update(+id, updateIntegrationDto);
+  // }
+
+  // @Delete(':id')
+  // remove(@Param('id') id: string) {
+  //   return this.integrationsService.remove(+id);
+  // }
+}
